GM2 - Collectibles
Made star collectible, added score ui to hud, and adjusted hud instantiation to allow generic creation. There is an array of types that can be updated to configure hud widgets. BI_PlayerUI can be added to a widget so it it initialized at start with a reference to the player.