Camera Fixes and Win Screen
Implemented final camera fixes for this iteration. Also implemented a simple win screen. Didn't want to get too fancy with it since core will likely come up with that later, so it just displays who won and with how many points. Had to make changes to all of the hardpoint level BPs, as a bug there was keeping the game from ending correctly. (TL;DR: did Sai's work for him)