AOE Hazard Tweaks
Made a bunch of changes to the AOE Hazard and Hazard Spawner. Mostly just refactoring code to make it legible and making some macros and functions for commonly used functionalities. Had to change how the targeting was done, as the old targeting system could cause drift out of the arena over time. Added a DA for the hazard spawner that still needs to be setup. (TL;DR: did Sai's work for him)