Generalized Hazards
Changed up the hazards to be more general. Hazard should now be used as a parent class. Override "Impact" and "Set Size and Speed" functions to create new kinds of hazard. Also finished implementing the Hazard Spawner PDA, so it should be easier for design to prototype with. (TL;DR: did Sai's work for him)