Comet Hazards
Added in falling hazards that drop ice on the ground where they impact. Currently drops meshes because I might be slightly brain dead when it comes to materials, but I will try to get it working with decals next. Also continued to pare down the hazard system for readability (TL;DR: did Sai's work for him)