Ice Decals
Changed the ice hazards to use decals instead of a mesh with the ice material. There still has to be an invisible mesh to hold the physics material, but the visuals are now all decal-based. Poached a mask for the decal from a circle texture in existing assets -- may want a dedicated mask from art in the future. (TL;DR: did Sai's work for him)