Crown Slow & Updates
The player now slows down while the crown is held. Included options for a continuous slow or for incremental step slow when the player scores. Also made some minor changes that were discussed such as increasing collision size and setting the crown to follow a socket. (TL;DR: did Sai's work for him)