Star Steal
This is placed in the fast to slow custom collision checks.

A variable for the star stealing amount has been exposed on spawn and set to instance editable. This variable affects how many stars the stealing player gains, and how many stars the other player will have stolen.

Uses a custom event which is similar to "Event Lose a star", but within the same event it adds a star score to the star stealing player by connecting the target pin to self. For the player that must have a star stolen to lose a star in, the cast node is connected to the event Star Steal's target pin to work.