Working Collision Check & Responses
The reason the custom collision code kept ticking was because speeding collision is activated throughout the charge.

The fix was to get the "Repulser Collision" node and leave the "toggle" bool unchecked. With a delay of 0.2 seconds, speeding collision's overlap event performs the logic in a working manner.

For this fix there was no need to modify the previous code in the speeding collision overlap. It only needed the repulse collision node.
Will continue work on this.