Working Fast v Fast Collision Check
There's now a bool function check for when the player's speeding collision component is active so that both player's can give each other an impulse.

One bug, is if player's begin charging when they are close enough for their speeding collision components to overlap, it does a fast vs slow impulse instead of fast vs fast.
Will continue to work on this